Extract rankAffixes
